+++ app/globals.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap');
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -95,3 +93,4 @@ html {
 .animate-glow {
   animation: glow 3s ease-in-out infinite;
 }
+
