+++ components/features/TryOurAI.tsx
@@ -5,58 +5,58 @@ const TryOurAI: React.FC = () => {
     <div className="fixed bottom-[110px] left-8 z-30 pointer-events-none">
       {/* Handwritten Text */}
       <div className="relative">
-        <svg 
-          width="120" 
-          height="120" 
+        <svg
+          width="120"
+          height="120"
           viewBox="0 0 120 100"  // Expanded viewBox for better arrow
           className="text-navy-900"
           style={{ transform: 'rotate(-5deg)' }}
         >
           {/* Try */}
-          <text 
-            x="4" 
-            y="20" 
-            className="fill-current text-blue-900" 
-            style={{ 
-              fontFamily: 'Lucida Calligraphy, Edwardian Script ITC, cursive', 
-              fontSize: '22px', 
+          <text
+            x="4"
+            y="20"
+            className="fill-current text-blue-900"
+            style={{
+              fontFamily: 'Lucida Calligraphy, Edwardian Script ITC, cursive',
+              fontSize: '22px',
               fontWeight: 'normal',
               fontStyle: 'italic'
             }}
           >
             Try
           </text>
-          
+
           {/* Our */}
-          <text 
-            x="58" 
-            y="30" 
-            className="fill-current text-blue-900" 
-            style={{ 
-              fontFamily: 'Lucida Calligraphy, Edwardian Script ITC, cursive', 
-              fontSize: '22px', 
+          <text
+            x="58"
+            y="30"
+            className="fill-current text-blue-900"
+            style={{
+              fontFamily: 'Lucida Calligraphy, Edwardian Script ITC, cursive',
+              fontSize: '22px',
               fontWeight: 'normal',
               fontStyle: 'italic'
             }}
           >
             Our
           </text>
-          
+
           {/* AI */}
-          <text 
-            x="25" 
-            y="55" 
-            className="fill-current text-blue-900" 
-            style={{ 
-              fontFamily: 'Lucida Calligraphy, Edwardian Script ITC, cursive', 
-              fontSize: '22px', 
+          <text
+            x="25"
+            y="55"
+            className="fill-current text-blue-900"
+            style={{
+              fontFamily: 'Lucida Calligraphy, Edwardian Script ITC, cursive',
+              fontSize: '22px',
               fontWeight: 'normal',
               fontStyle: 'italic'
             }}
           >
             AI
           </text>
-          
+
           {/* Curved Arrow - smooth downward curve */}
           <path
             d="M 65 52 Q 80 65 85 80"
@@ -66,7 +66,7 @@ const TryOurAI: React.FC = () => {
             className="text-blue-900"
             strokeLinecap="round"
           />
-          
+
           {/* Arrow Head - filled triangle pointing down */}
           <path
             d="M 85 80 L 80 75 L 90 75 Z"
@@ -80,3 +80,4 @@ const TryOurAI: React.FC = () => {
 };
 
 export default TryOurAI;
+
