+++ components/layout/Logo.tsx
@@ -19,3 +19,4 @@ const Logo: React.FC = () => {
 };
 
 export default Logo;
+
