+++ components/sections/Hero.tsx
@@ -107,3 +107,4 @@ const Hero: React.FC = () => {
 };
 
 export default Hero;
+
