--- pages/_app.tsx
@@ -1,7 +1,0 @@
-import type { AppProps } from 'next/app';
-import '../src/index.css';
-
-export default function App({ Component, pageProps }: AppProps) {
-    return <Component {...pageProps} />
-}
-
