--- pages/_document.tsx
@@ -1,24 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document';
-
-export default function Document() {
-    return (
-        <Html lang="en">
-            <Head>
-                <link rel="icon" type="image/svg+xml" href="/favicon.ico" />
-                <meta
-                    name="description"
-                    content="Levrok Labs - AI Agency engineering intelligent solutions for tomorrow's challenges"
-                />
-                <link
-                    href="https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap"
-                    rel="stylesheet"
-                />
-            </Head>
-            <body>
-                <Main />
-                <NextScript />
-            </body>
-        </Html>
-    );
-}
-
